requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.6.5
python-dotenv>=1.0.1
python-multipart>=0.0.9
httpx>=0.27.0
websockets>=12.0
pyyaml>=6.0
litellm>=1.40.0
openai>=1.30.0
unidiff>=0.7.5
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
arxiv>=2.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
# Optional: docker>=6.1.3 chromadb>=0.4.24 gitpython>=3.1.40 watchdog>=3.0.0
